requirements/jenkins.txt
-r include.txt

django-jenkins==0.12.1
coverage==3.5.2
pylint==0.25.1
pep8==1.0.1
pyflakes==0.5.0
